:recycle: moves FileCentricDocumentField to common place
