+++ components/pages/file-repository/FileTable/index.tsx
@@ -25,7 +25,6 @@ import {
   FileRepositoryTableQueryData,
   FileRepositoryTableQueryVariables,
   FileRepositoryRecordSort,
-  FileCentricDocumentField,
   FileRepositoryRecordSortOrder,
   FileRepositorySortingRule,
 } from './types';
@@ -46,6 +45,7 @@ import { SortedChangeFunction } from 'react-table';
 import { useTheme } from 'uikit/ThemeProvider';
 import TsvDownloadButton from './TsvDownloadButton';
 import useFileCentricFieldDisplayName from '../hooks/useFileCentricFieldDisplayName';
+import { FileCentricDocumentField } from '../types';
 
 const DEFAULT_PAGE_SIZE = 20;
 const DEFAULT_PAGE_OFFSET = 0;
