+++ components/pages/file-repository/types.ts
@@ -1,0 +1,12 @@
+export enum FileCentricDocumentField {
+  object_id = 'object_id',
+  'donors.donor_id' = 'donors.donor_id',
+  study_id = 'study_id',
+  data_type = 'data_type',
+  file_type = 'file_type',
+  'analysis.experiment.experimental_strategy' = 'analysis.experiment.experimental_strategy',
+  'file.size' = 'file.size',
+  'donors.gender' = 'donors.gender',
+  variant_class = 'variant_class',
+  file_access = 'file_access',
+}
